API call demo using retrofit

+++ app/src/main/java/com/example/kb/clearsky/model/api_specific/WeatherInfo.java
@@ -35,7 +35,7 @@ public class WeatherInfo {
 
     @SerializedName("temp_kf")
     @Expose
-    private Integer tempKf;
+    private Double tempKf;
 
     public Double getTemp() {
         return temp;
@@ -93,11 +93,11 @@ public class WeatherInfo {
         this.humidityPercent = humidityPercent;
     }
 
-    public Integer getTempKf() {
+    public Double getTempKf() {
         return tempKf;
     }
 
-    public void setTempKf(Integer tempKf) {
+    public void setTempKf(Double tempKf) {
         this.tempKf = tempKf;
     }
